+++ plugins/azure-wakamiti-plugin/CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog][1],
 and this project adheres to [Semantic Versioning][2].
 
 
+## [unreleased]
+
+### Changed
+- The identifier is now incorporated into the name of the test case, in the 
+  format `[ID-XX] Scenario Name`.
+
+
 ## [2.1.2] - 2025-02-26
 
 ### Fixed